+++ src/main/java/cech12/usefulhats/item/IAttackTargetChanger.java
@@ -2,9 +2,10 @@ package cech12.usefulhats.item;
 
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 
 public interface IAttackTargetChanger {
 
-    void onLivingSetAttackTarget(Mob entity, Player targetPlayer);
+    boolean avoidMobChangingTarget(ItemStack itemStack, Mob entity, Player targetPlayer);
 
 }