+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.19.2-3.0.1.0] - 2022-09-13
+### Changed
+- Update mod to Forge 1.19.2-43.1.23
+- Make Halo related to the new boss tag of Forge
+- optimize Halo damaging to be more precise
+
+### Removed
+- removed Halo detecting range config option, because it is not needed any longer
+
 ## [1.19-3.0.0.0] - 2022-07-18
 ### Changed
 - Update mod to Forge 1.19-41.0.96